backend/requirements.txt
fastapi
uvicorn[standard]
pandas
python-dotenv
requests
pymongo
pypdf
twilio
python-multipart
